chore(crudviews): Added formMarkup Rerplaced model with formValues, added commitzen
formMarkup property  was added to all the components that used formPath in which the property was
missin to allow the user the ability to pass the markup if available. On the GenericDetailsPage the
model was replacing the formValues when formValues, if passed, was supposed to override the model.
That issue has been addressed. Also made the repo commitzen friendly