fix(layout): Removed references ro uuid and replaced with nanoid/non-secure to fix crypto issue
When using persisted layout there's an issue that is caused by uuid module to crypto. The browser complains about crypo not being defined because, somehow, the registration of the layout happens before crypto is available. The error is https://github.com/uuidjs/uuid#getrandomvalues-not-supported. So, I resorted to using non-secure nanoid because, though predictable, it doesn't use crypto